RMLNQ-67 Replace ExpressionVisitor in .NET 4.0 with native implementation.
In .NET 4.0, ExtensionExpressions and custom ( = unknown) Expressions are both handled by the VisitExtension method. Unknown Expressions must either be reducible or implement the VisitChildren() method. Otherwise, an exception will be thrown. Note, that VisitChildren() can return 'this' as the most basic implementation.

In .NET 3.5, ExpressionVisitort.VisitBinary() has been changed to call VisitAndConvert() instead of Visit() for the Conversion-Expression. This removes the call to ExpressionVisitor.Visit() when the Conversion-Expression is null.

In .NET 3.5, ExpressionVisitor.Visit(ReadOnlyCollection, visit) has been changed to support the visit method returning null-values.